Create quickstart_final.html
Part 1 of Assignment 1. "Create a Leaflet map and share it on GitHub of 5 places of your interest in St. Louis. Include popups! If you're up to it, change the symbology for the points.
In quickstart_final.html, add 5 markers.
Then, bind a Popup to each one:
Remove other markers, such as circles and polygons."